feat: add users already exists verifcation in seed

+++ src/services/database/seeders/index.ts
@@ -4,6 +4,13 @@ import { prisma } from "../prisma.client";
 
 async function seedDatabase() {
 	console.log("🌱 Iniciando seed do banco de dados...");
+	const usersExists = await prisma.user.findMany();
+	if (usersExists.length > 0) {
+		console.log(
+			"❌ Usuários já existem no banco de dados. Abortando seed.",
+		);
+		return;
+	}
 
 	await prisma.user.deleteMany();
 	await prisma.account.deleteMany();
